Create courses1.2
includes the updated courses table with undergrad CSCE courses, with the new 10 columns. Status is to be either empty, null or 'open' for all courses, as there won't be any implementation for it. The last column is just a unique index column to make it easy to modify the table later.