TEST COURSE DATABASE
You can use this to test your code when you're attempting to connect to a database for values. Save this as 'course.sql' and with xampp running Apache and MySQL, go to PHPMyAdmin (by clicking the Admin button for MySQL in the XAMPP Control Panel or by going to www.phpmyadmin.co). At the top of the page, click Import, and select course.sql to import. After the database is loaded, try connecting to it using the sample connection file I uploaded just before this. Just change the lines

$connection = mysql_connect('sql2.freemysqlhosting.net','sql260064','nY6%eA9!') or die(mysql_error());
$database = mysql_select_db("database1",$connection);

to

$connection = mysql_connect('localhost','root','') or die(mysql_error());
$database = mysql_select_db("database1",$connection);


If you get an error while importing or connecting, let me know. If you have XAMPP set up with Apache and MySQL running, you should just be able to use this database to test all of your code that needs to work with the courses database.